meta-lxatac-software: rauc: add a date to the bundle version
RAUC allows us to query the bundle version string of remote bundles over dbus.
This means we can pass a bundle URL to RAUC and it will respond with the bundle
version string.
If we now include a version and date in the version string we can implement
upgrade notifications based on it, á la "There is an update bundle available
that is 30 days newer than your current version. Do you want to upgrade?",
while only requiring new bundles to be placed on a known location on a dumb
http server.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>