meta-lxatac-software: tacd: make sure the DUT is powered off on exit
The polarity of the DUT Power Enable and DUT Discharge lines has changed
between the -R01 and -R02 revisions of the power board, as -R01 boards had an
issue with the power being on before the first write was sent to the GPIO
expander chip.

The change was reflected in tacd itself but forgotten in the service file.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>