meta-lxatac-bsp: linux-lxatac: enable LCD driver and disable OLED driver
This has the added benefit of no longer requiring anything from staging which
can thus be disabled.

The panel-mipi-dbi driver used for the LXA TAC LCD uses a small "firmware file"
(a list of commands to send to the LCD on init) to bring up the display.
Include the firmware in the kernel image so we do not have to build the driver
as module and have the LCD set up very early in the boot process.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>