meta-lxatac-software: rauc bundle: hook.sh fix lxa-iobus cache migration
The lxa-iobus server maintains a cache of IOBus node IDs it has seen before
to speed up re-discovery of said nodes (known IDs are tried befor performing
a full LSS scan).

This cache is stored in /var/cache/lxa-iobus/lss-cache.

The migrate function in hook.sh only works on files, so it will not currently
migrate the lss-cache inside of /var/cache/lxa-iobus.
Fix that by pointing to the file directly. Also make sure the directory to
migrate to exists on the system.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>