meta-lxatac-{software,bsp}: add licenses to helper recipes
We have a few recipes that copy configuration from A to B or provide
helper scripts on the TAC.
We set them to a CLOSED license before we knew how, where and under which
license we would publish meta-lxatac.
Now we know that and need to set actual licenses for them.
MIT sounds like a sane choice in this case.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>