meta-lxatac-software: python3-canopen: remove recipe
This recipe was added in 2020 in a commit titled "software:
add python3-canopen" to allow the subsequent addition of a predecessor of
what now is the lxa-iobus-server.

Time has gone by, the lxa-iobus-server has changed name, location, got
a recipe in meta-labgrid and some time in between stopped using
python3-canopen, but this recipe has stuck around with us as a monument
to that time.

Now its time to go has come.

Bye python3-canopen_1.1.0.bb

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>