meta-lxatac-software: chrony: use Linux Automation GmbH ntp pool zone
The ntp pool project requires vendors to apply for a vendor zone[1] so that
they can detect miss-behaving clients and contact the responsible parties,
or ask heavy users for a contribution.

There is also a note about encouraging users to join the pool in the
configuration file:

    If you have an open source ntpd implementation or an operating system
    including ntpd, we ask that you make a reference in the configuration
    file or documentation encouraging people to join the pool.

which I have also added.

[1]: https://www.pool.ntp.org/vendors.html

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>