labgrid-exporter: export FTDIs with multiple separate interfaces
Some FTDIs pop up as multiple USB interfaces on the host system.
Let's export there as well. This could've been done in the userconfig.yaml,
but it's more convenient to just do this here centrally.

We'll just export the main interface and three more and assume 4 FTDIs
is enough for everybody.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>