meta-lxatac-software: rauc bundle: hook.sh migrate authorized_keys
Currenty we recommend installing custom SSH keys to
/home/root/.ssh/authorized_keys for custom images.
This does however mean that we can not migrate said file in the RAUC hook,
which in turn means locking out users of the "vanilla" bundle on every upgrade.
This is not ideal.

Change the advise in the README to installation of keys from bundle to /etc
and start migrating /home/root/.ssh/authorized_keys.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>