meta-lxatac-bsp: rauc: add release certificates and insecure devel cert
The testing and stable certificates are used to verify official bundles
provided by the Linux Automation GmbH.

The devel certificate is provided along with its private key making it
insecure and unsuitable for production use.
The devel certificate is provided so that LXA TAC RAUC bundles can be
built without having to generate keys and certificates first.

When installing from a RAUC bundle only the certificate said bundle was
signed with will be activated in the installed system.

On non-customized _images_ the devel certificate will be activated.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>