meta-lxatac-bsp: lxatac-repart: use /system-update for a clean repart-boot
If a /system-update file/symlink exists systemd will start
the system-update.target instead of the default target.

After the system-update.target is reached the /system-update file is removed
and the system rebooted, resulting in a normal boot the next time around.

By not starting any services but systemd-repart in system-update.target we
can make sure that the /srv directory stays empty on the first boot on which
the partition for it is created.

On the second boot the /srv partition will already be created and ready for
mounting.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>