meta-lxatac-bsp: systemd: move journal flush to multi-user.target
We use system-update.target to re-partition the disk on first boot,
meaning we do not have all partitions available on the very first boot
and need to keep the mountpoints where they should be mounted clean.

This breaks a dependency chain that would result in a /srv/journal
directory being created in the rootfs's /srv:
system-update.target -> sysinit.target -> systemd-journal-flush.service
-> var-log-journal.mount

This can not be done by modifying any WantedBy= or similar, as systemd
just does a hard install of this symlink.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>