meta-lxatac-software: recipes-support: add recipe for i.MX uuu
NXP Mfgtools uuu is the tool commonly used to upload U-Boot on i.MX
SoCs. We already ship imx-usb-loader for uploading barebox to i.MX SoCs,
so in preparation for supporting upload of U-Boot, let's build uuu
as well. The recipe is based on the version in meta-freescale[1], but
instead of git, it downloads source tarballs from the Github releases
page. 1.5.21 is currently the latest release that's not tagged as
prerelease.

[1]: https://git.yoctoproject.org/meta-freescale/commit/recipes-devtools/uuu/uuu_git.bb?id=c77a3e2391f

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>