poky/meta-*: update to nanbield/master branches
The libiio version in meta-oe was updated and we no longer have to carry the
backported patch.

Likewise with the dt-utils patch and meta-rauc.

Our ".rootfs.ext4" -> ".ext4" workarounds in lxatac-core-bundle-base.bb and
emmc-image/lxatac/genimage.config are also no longer required.

The perf package now handles the libtraceevent dependency correctly and we
do not need that bbappend anymore.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>