emmc-image: remove size workaround
genimage knows about the 'size' option for images to fill the image to a
certain size, and since genimage v14 it supports the 'fill' option to
write all zeroes at the end of the image and not cut them off to
optimise image size. Use this to blow up the images to the full 2 GiB,
and remove our invisible-partition-at-the-end hack.

Since the GPT header and the protective MBR occupy the first 1 MiB (with
1 MiB alignment as specified), and the last 1 MiB at the end of the
image is occupied by the GPT backup header, and there is a 1 MiB
reserved partition, the rootfs must now be 3 MiB less than 2 GiB to
fit into the image.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>