meta-lxatac-software: github-act-runner: add recipe
The official GitHub action runner client is written in C# and uses
(as far as I could figure out) the non-free(?) part of the .NET
Framework, making it hard to build.

This alternative client is written in Go and relatively easy to build.
We'll have to see how compatible it is in practice.

The need for nodejs on the TAC is due to all "uses:" steps for
runners being implemented in javascript (*sigh*).

To configure a runner you have to into the Actions setting of your GitHub
repository, and add a new self-hosted runner.
After configuring the OS and Architecture (Linux / ARM) you can use the
presented token to register the runner:

  $ mkdir -p /etc/github-act-runner
  $ cd /etc/github-act-runner
  $ github-act-runner configure \
      --name $(uname -n)
      --token 01234567899876543210012345678
      --labels dut-raspberrypi
      --url https://github.com/user/repository

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>