Revert "emmc-image: remove size workaround"
This change did not do quite what we wanted it to do.
We do actually want the two rootfs partitions to be 2G in size,
not the whole image.

Getting rid of the "partition end" hack is actually something
we want, so we will re-do that in another commit but a bit differently.

This reverts commit 9b5bf5b79aeb63161435918fdb9ad4da8614b6ee.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>