meta-lxatac-bsp: emmc-image: do not fill and write GPT backup header
A GPT parition table places an additional header at the _end_ of the device
it is on. Our eMMC is however 16Gb large and the image we initially deploy
is much smaller than that.
The GPT header genimage generates for us does thus not land at the end of
the eMMC and every tool looking for it there will not find it.
Writing it is thus quite useless.

Removing it means that the image now longer has to span the whole 2Gb we
previously had. It can instead end right after the file system.
To make the image end right after the file system we also need to remove the
dummy partition placed at the end of the image just to force padding of the
image from the end of the file system to the dummy partition.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>