meta-lxatac-bsp: emmc-image: clear barebox state on install
Previously the range in the emmc sparse image corresponding to the
barebox state was not written and thus considered sparse.
This meant that it was not re-written when writing the sparse image.

Explicitly write zeros to the state partition to clear it.

Note that this change removes the alignment parameter from the image
because the new partitions do not align to a 1M boundary.
This means we need to specify an explicit offset for each partition
from now on.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>