meta-lxatac-bsp: barebox: use partition labels to find root partitions
This change is inspired by a similar DistroKit commit [1].
It allows us to change the partition numbering without having to adjust
the partition numbers in barebox.

Adjusting partition numbers in barebox is dangerous because the
numbering is set on the first boot after a new install from an image
and does not update on subsequent RAUC installs.
This means there can be TACs with different partition layouts in the
wild and fixing that requires a re-install from image.

Just using partition labels works around that problem.

[1]: https://git.pengutronix.de/cgit/DistroKit/commit/?h=next&id=c4a597c6eed61b0c34e39727f63d17fd8b96c947

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>