meta-lxatac-bsp: linux-lxatac: upgrade to 6.7
umpf tag is now: 6.7/customers/lxa/lxatac/20240108-1.
The tag includes the following topic branches:

 - v6.7/customers/lxa/lxatac
 - v6.7/topic/pwm-stm32
 - v6.7/topic/pwm-backlight
 - v6.7/topic/tty-trigger

No interesting changes this time around. The KSZ drive strength
HACK got even HACKier because the #defines are no longer in scope,
so I replaced them with the computed value.
We only need it until the drivestrength setting is fixed mainline
so that it is applied after every reset of the IC (of which there
are some in the initialization sequence of the driver).

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>