meta-lxatac-software: distro: tacos: add TAC OS distro based on meta-ptx
Up until now we did not need to track any differences from meta-ptx's
sample distribution (or we we put them into e.g. the machine, even though
that's not technically correct).

That changes in one of the follow-up commits in which we set static uid
values for all users, which is a distro setting and not a machine setting.

So please give a warm welcome to TAC OS - The LXA TAC OS.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>