README: enable an update channel after install from image
The releases in the stable and testing channel on
`downloads.linux-automation.com` are byte-for-byte equivalent,
except for the RAUC bundles, which are re-signed before they are uploaded.
When installing a testing or stable bundle the update channel to enable in
the new slot is chosen based on the certificate the bundle was signed with.

For a from-scratch installation we do not have this information,
so the user has to choose a channel to enable.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>