meta-lxatac-software: tacd: failsafe: gpiofind no longer exists
The gpiofind command was removed, so the tacd-failsafe command would
currently fail (which is kind of ironic).

Instead gpioset can be directly supplied with the GPIO names.
The GPIOs are not guaranteed to keep the assigned value once gpioset
exits, so we just keep it running for some time before trying to restart
the tacd.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>