CI: fix distribution version consistency checks
To follow a well honored tradition our first CI job we added failed
instantly when it landed in the main branch.
Now that the traditions have been honored we can fix it.

There were multiple things wrong with the previous version regarding
CI jobs running for a branch or tag (instead of jobs running for a pull
request):

  - The GITHUB_REF variable contains a name like "refs/heads/mickledore",
    not just "mickledore". We want GITHUB_REF_NAME.
  - The GITHUB_BASE_REF variable is also set for non pull request jobs,
    but it is set to an empy string. Check for that to decide if we are
    running in "pull request mode".
  - When running for a tag the branch name check for the os codename
    does not make sense. Skip it in that case.

Fixes: 7fa11b2 ("CI: add distribution version consistency checks")
Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>