meta-lxatac-bsp: libtraceevent: backport "build with Meson" from oe-core
This fixes an issue where the libtraceevent build would sometimes
generate .so files with some objects missing due to a race condition
in the Makefile [1].
One of these builds made it into the sstate of our CI builds,
making them fail.

The upstream fix is to just move to meson as a build system,
so lets just do the same.

This only affects our mickledore branch, as the backport comes from
the nanbield poky branch (and langdale did not use libtraceevent).

Thanks to Enrico for finding the respective oe-core patch!

[1]: https://git.yoctoproject.org/poky/commit/?id=142b6921bde5b22bcd7e0e5e8d895575d0fb6c64

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>