meta-lxatac-bsp: machine: lxatac: use RAUC releases instead of git version
Currently the rauc version in rauc-1.11.1.bb is newer than the version
in rauc_git.bb. This is obviously an uncommon situation, but it made
us aware that we set RAUC_USE_DEVEL_VERSION=1 without a clear reason
or need.

Released RAUC has all the features we need. Use it.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>