meta-lxatac-bsp: rauc: add patch to fix RAUC info / update notifications
Currently the update notifications on the LXA TAC LCD and web interface
do not work because the tacd can not get informations about the RAUC
bundle on the update server via RAUC's DBus API.

This is because the process that RAUC spawns to handle the nbd bridge
for http streaming crashes due to an unitialized pointer access.

This patch is taken from RAUC pull request 1332 [1], which fixes the issue.

[1]: https://github.com/rauc/rauc/pull/1332

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>