meta-lxatac-software/containers: Update to debian:latest
This effectively bumps the Debian version to the most recent release at
that time.
This way we will not need to update the script with every Debian
release.

This also aligns the behavior with `container-start.sh` that simply
pulls `debian` if no container is present.
And this also defaults to `latest`.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>