meta-lxatac-bsp: barebox: enable fastboot oem command
This allows us to run barebox commands without the need for a serial
console (using the dedicated serial cable or the barebox CDC ACM gadget).

This solves an issue where the TAC would just continue auto booting
when trying to flash a new image onto the TAC via USB.

Also document in the README that we no longer need workarounds for
the watchdog and can now boot the flashed image without requiring a
power cycle.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>