meta-lxatac-bsp: rauc: follow meta-rauc and split rauc.bb and rauc-conf.bb
This reduces how often rauc has to be recompiled.
And it also makes the main rauc recipe machine-independent.
This does not help us too much because we only have one machine in this
BSP, but it should be nice for others.

Split our .bbappends into one for rauc itself and one for the config.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>