meta-lxatac-software: lxatac-core-bundle-base: migrate /etc/hostname
We set default hostnames based on the serial number of the LXA TAC,
(e.g. lxatac-00011) which is a sensible thing to do.

But users may want to set their own hostnames due to constraints of
their infrastructure or just personal preference.

Enable them to do so persistently across RAUC updates.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>