meta-lxatac-bsp: lxatac-factory-data: remove reliance on hostnamed
We want to pull the exectution of the factory-data script earlier in the boot
process, so it runs before udev starts.
Relying on a DBus service to get and set the hostname goes against that goal.
Just asking the kernel and writing to /etc/hostname is more or less equivalent
to what we have done before, so just do that.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>