meta-rauc: update to version with rauc v1.11.3
This bugfix release fixes an issue with system.confs that use symlinks,
which we do.
In most cases we are not affected by this bug because we use bootchooser,
so the correct slots are found via bootchooser kenel commandline parameters
but sometimes we also boot via e.g. `boot root-a` on the barebox
commandnline in which case the bug triggers and RAUC fails to start.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>