meta-lxatac-bsp: barebox: build binaries without STM32 header
Ahmad Fatoum has pointed out that the .stm32 images we currently use only
work by chance for your use case because the stm32 header in front is no
longer parsed by TF-A and is directly executed.

It seems that we were just lucky until that the bytes in the header
evaluates to arithmetic instructions that don't crash.

Apply a patch provided by Ahmad that builds the barebox binary without
STM32 header.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>