kernel: Add a bunch of serial fixes
When /dev/ttySTM1 is closed while there are still some characters
pending to be sent, the kernel dereferences a null pointer which locks
up the kernel.

This is fixed by two patches cherry-picked from next.

Another problem is that if the UART is closed while being throttled,
the irq handler thinks it's still throttled after reopen which results
in it not handling RX events resulting in a stuck irq starving the
machine. This is fixed by two further patches. The first only improves
the driver's behaviour generally on stuck irq, the second is the actual
fix.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>