meta-lxatac-software: python3-twisted: backport fix for empty packages
This is a backport of meta-oe master commit bd1b5cde3
("python3-twisted: prepend split PACKAGES"),
which fixes an issue where the `python3-twisted` package would gobble up
all of the files in the split stage (because it matches broadly and comes
before the other packages in the list) and most of the other packages
ended up empty.
Since this workaround is only required until that commit makes it into
the scarthgap branch go the easy route and just hardcode the package
order achieved by bd1b5cde3.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>