meta-lxatac-software: py-spy: drop recipe
While py-spy is a useful tool it has sadly not seen a new release in some
time, which results in it no longer being compatible with the python
version we ship on the LXA TAC.
This means it does not make a lot of sense to ship py-spy right now.

We can re-enable the recipe once py-spy is updated to support current
python versions.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>