meta-labgrid: update to a version that adds grpc support
This marks a breaking change compared to the previous release version of
labgrid, as this new development version now uses grpc instead of crossbar
to communicate between the client, coordinator and exporters.

This means all of these components need to be updated to the grpc-based
versions at the same time.

See https://github.com/labgrid-project/labgrid/pull/1469 for more
information about this change.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>