meta-labgrid: use scarthgap branch instead of master
The tip of the master branch is no longer compatible with the scarthgap
release, since it has the UNPACKDIR changes that are required for styhead.
Since the scarthgap branch exists now and is up to date
(it didn't initially, which is why this was set to "master") use that now.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>