ci: build: harmonize the build job with the ones used in other projects
We have since rolled out out the forrest runner for other projects
(meta-ptx, meta-rauc and meta-labgrid) and have decided to use shared
sstate, git and downloads directories that are updated via rsync at the
end of runs.

This makes the meta-lxatac project a special case and the only one that
persists disk images instead.

Remove this special-casing by also employing the sstate rsyncing logic
for meta-lxatac.

The other projects take extra care to enable running the build process
on public GitHub runners by using an sstate mirror and using that in the
public runner case.
But meta-lxatac's heavy bones make it unlikely that a build will succeed,
event with an sstate mirror, so this extra logic has been left out.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>