ci: build: upload the RAUC bundle at the end of the build
We do not want to be wasteful with the upload traffic we consume,
hence why we did not upload any artifacts previously.
(The sparse image for the eMMC is ~840Mb and the RAUC bundle ~290Mb).

Being able to test the the generated artifacts would however be super
convenient, so upload at least the RAUC bundle for now.

The upload step is configured to upload the file that has the build
date and time in its filename and not the symlink linking to it.

(The upper one in this listing:)

  $ ls -ahl lxatac-core-bundle-base-*
  -rw-r--r-- 1 lgo lgo 299M Aug 30 13:02 lxatac-core-bundle-base-lxatac-20240830102608.raucb
  lrwxrwxrwx 1 lgo lgo   51 Aug 30 13:02 lxatac-core-bundle-base-lxatac.raucb -> lxatac-core-bundle-base-lxatac-20240830102608.raucb

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>