meta-lxatac-bsp: linux-lxatac: add workaround for broken fbdev emulation
Starting with commit 86634fa4e6aef ("Merge v6.10-rc6 into drm-next") the
tacd fails to write to the /dev/fb0 device with a SIGBUS error.

This merge commit is the first affected commit because it brings together
these two commits for the first time:

- commit 9317ff69170d3 ("drm/tiny/panel-mipi-dbi: Use fbdev-dma")
- commit d92a7580392ad ("drm/fbdev-dma: Only set smem_start is enable per
   module option")

The first one changes panel-mipi-dbi from being a `drm_fbdev_generic`
device to being a `drm_fbdev_dma` device.
The second one prevents `drm_fbdev_dma` devices from leaking the memory
address of the framebuffer to userspace unless the
`drm_kms_helper.drm_leak_fbdev_smem=1` kernel commandline option is set.

This being an issue is likely a bug in the panel-mipi-dbi driver that
should be fixed there, but this workaround makes sure we can use 6.11
until the driver is fixed.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>