ci: build: enable artifact upload to the local Forrest runner
This makes it so that artifacts are no longer uploaded via the internet to
GitHub's cloud infrastructure but instead stay on the physical server the
builds run on.

Since the built images will mostly be consumed from within the building the
server is located in this should reduce the upload bandwidth used,
allowing us to also upload the eMMC image and not only the RAUC bundle.

Since the uploaded files are served from a webserver as-is,
instead of being packaged in a zip file like "normal" artifacts are,
they can also be used in a `rauc install` call directly,
making testing of a build much easier.

The artifact file that _is_ uploaded to GitHub's infrastructure contains
`.desktop` files with `Link=https://...` entries withing them that point
to the location of the actual file.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>