meta-lxatac-software: python3-labgrid: use SRCREV from meta-labgrid
Commit e3b9c26 should have updated the labgrid version used to the one
configured in meta-labgrid.
Instead it only updated the recipe (and dropped the dependency for
autobahn etc.), so we currently have a labgrid version installed that
still depends on autobahn, but with dependencies for a grpc based labgrid
version.
This results in `labgrid-exporter.service` failing to start:

    Traceback (most recent call last):
      File "/usr/bin/labgrid-exporter", line 5, in <module>
        from labgrid.remote.exporter import main
      File "/usr/lib/python3.12/site-packages/labgrid/remote/exporter.py", line 19, in <module>
        from autobahn.asyncio.wamp import ApplicationRunner, ApplicationSession
    ModuleNotFoundError: No module named 'autobahn'

Fix that by _actually_ using the SRCREV from meta-labgrid.

Fixes: e3b9c26 ("meta-labgrid: update to a version that adds grpc support")
Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>