meta-lxatac-software: python3-labgrid: use release v24.0.2
Instead of some more or less random master commit.

The previous labgrid commit, when used with the ser2net version on the TAC
results in the following error when trying to connect to a console via
labgrid:

> Invalid accepter port name/number 'telnet(rfc2217,mode=server),48431':
> Invalid data to parameter

This is because the rootfs contains ser2net 4.6.1, which requires,
for our purposes, that the command line contains `,tcp`.

But the labgrid commit installed on the target does not add this argument.

Update the commit hash to the one for the v24.0.2 release.

This is fixed in the scarthgap branch by commit 6876eb8
("meta-lxatac-software: python3-labgrid: use SRCREV from meta-labgrid")
which removes the SRCREV line altogether and instead uses the labgrid
version from meta-labgrid.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>