meta-lxatac-bsp: rauc: add placeholder for polling section to system.conf
The polling information (interval, URL, auto-install behaviour, etc.)
is only known to the tacd. To pass them to RAUC the tacd will read the
config in `/usr/lib/rauc/system.conf` replace the placeholder added
in this commit with a proper polling secion and place the result
in `/run/rauc/system.conf`. After a deamon restart RAUC can handle
update polling and installation on its own.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>