meta-lxatac-bsp: linux-lxatac: disable unused VT options
We don't use the VT nor the PTYS support, therefore disable it.

Same is true for line-descipline drivers on-top of a serial device. The
line-descipline is replaced by serdev.

While on it, disable the rpmsg-tty driver support too since we don't use
the Cortex-M4 µC actually.

Disabling the DUMMY_CONSOLE is a side effect and is safe to do so.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>