meta-lxatac-software: lxatac-profile: set LG_COORDINATOR not LG_CROSSBAR
With the switch from crossbar to grpc the LG_CROSSBAR environment variable
has been replaced with LG_COORDINATOR.

Set the correct envrionment variable to enable local use of labgrid
commands.

There are currently two places where we use `LABGRID_COORDINATOR_IP`
and `LABGRID_COORDINATOR_PORT`: in this script (to set up
the `LABGRID_COORDINATOR` for interactive shell sessions) and
in `labgrid-exporter.servcie`.

In both places we just stick them together with a `:` and do not use the
split form at all.

We could thus simplify booth users by just having the concatenated form
in `/etc/labgrid/environment`, but we do not do that because this file
is migrated between RAUC installs and we want to maintain
backward-compatibility.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>