meta-lxatac-software: python3-labgrid: install config for git and release
We recently switched to a release version of labgrid instead of a git
version, but did not update the bbappend file name accordingly,
resulting in our custom config not getting installed anymore.

Fix that and make sure the config is installed for git versions and
release versions, should we switch again in the future.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>