meta-lxatac-software: add new Qualcomm Downloader (QDL) recipe
qdl is tool that communicates with Qualcomm SoCs to upload new software.
We don't have support for it in Labgrid yet, but we need to start
somewhere, so let's make it available for command-line usage outside of
Labgrid.

The recipe is taken from meta-qcom[1] with a change to PV to have it
reference the latest release instead of 0.0[2].
The recipe there points at the current master revision.
The project has releases, but the newest release is too old for the
RB3 Gen2 I intend to use this for.

[1]: https://github.com/qualcomm-linux/meta-qcom/commit/6b64c34ee71b45d
[2]: https://github.com/qualcomm-linux/meta-qcom/pull/930

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>