ci: build: also upload sstate for dependabot builds
This effectively reverts commit 50d1004 ("ci: build: only persist disk
images for protected branches"), because it causes us to build a lot of
code twice whenever dependabot brings in a poky update.

With this change the build triggered by the dependabot PR will already
upload it's sstate, meaning we do not have to rebuild from scratch once
the PR is merged.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>