ci: build: increase the build timeout to twelve hours
With the upgrade wo walnascar the build times have increased again to a
point where poky updates quite often result in the build being canceled
after the default timeout of six hours, especially if other jobs run
in parallel on the runner host.

This is quite annoying because it means no sstate will be uploaded and
the next job will just time out again.

As a stopgap just increase the build timeout to twice what it was before.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>