meta-lxatac-bsp: barebox: backport fix for MAC address list fixup
There was an issue with the mainlined TLV support for MAC address lists,
where the MAC addresses that were already assigned to an interface were
not filtered out from the list of spare MAC addresses passed to the kernel
anymore (which was the behavior before mainlining).

This patche fixes that and ensures that the MAC address assignment is the
same as before switching to the mainlined TLV support.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>