meta-lxatac-software: bcu: update and remove libcurl un-vendoring patch
Recent bcu no longer needs libcurl, because its self-updating feature was
removed.

Remove the patch via bbappend until the upstream recipe is updated.

The recipe was upstreamed by us for use in meta-lxatac, this is why
meta-lxatac serves as a testing ground for changes we want to upstream and
why we have a `.bbappend` here in the first place.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>